apps: move brick and device set definitions into a new file
